Solucionado problema con las interfaces genericas
