+++ src/main/java/com/kerjox/rubrica5bloque5/services/AutoresService.java
@@ -1,45 +1,13 @@
 package com.kerjox.rubrica5bloque5.services;
 
 import com.kerjox.rubrica5bloque5.entities.Autor;
-import com.kerjox.rubrica5bloque5.repos.AutoresRepo;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
 import java.sql.Date;
 import java.util.List;
 
-@Service("autoresService")
-public class AutoresService implements DataBaseService<Autor> {
+public interface AutoresService extends DataBaseService<Autor> {
 
-	@Autowired
-	private AutoresRepo repo;
+	List<Autor> findByNombre(String nombre);
 
-	@Override
-	public void insert(Autor autor) {
-
-		repo.insert(autor);
-	}
-
-	@Override
-	public void delete(Integer id) {
-
-		repo.delete(id);
-	}
-
-	@Override
-	public List<Autor> findAll() {
-
-		return repo.findAll();
-	}
-
-	public List<Autor> findByNombre(String nombre) {
-
-		return repo.findByNombre(nombre);
-	}
-
-	public List<Autor> findByDate(Date fNac) {
-
-		return repo.findByDate(fNac);
-	}
-
+	List<Autor> findByDate(Date fNac);
 }
