+++ src/main/java/com/kerjox/rubrica5bloque5/services/LibroService.java
@@ -1,43 +1,13 @@
 package com.kerjox.rubrica5bloque5.services;
 
 import com.kerjox.rubrica5bloque5.entities.Libro;
-import com.kerjox.rubrica5bloque5.repos.LibroRepo;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-@Service("libroService")
-public class LibroService implements DataBaseService<Libro>{
+public interface LibroService extends DataBaseService<Libro> {
+	List<Libro> findByIsbn(String isbn);
 
-	@Autowired
-	private LibroRepo repo;
+	List<Libro> findByAutor(String nombreAutor);
 
-	@Override
-	public void insert(Libro libro) {
-		repo.insert(libro);
-	}
-
-	@Override
-	public void delete(Integer id) {
-
-	}
-  @Override
-	public List<Libro> findAll() {
-		return repo.findAll();
-	}
-
-	public List<Libro> findByIsbn(String isbn) {
-
-		return repo.getLibroByIsbn(isbn);
-	}
-
-	public List<Libro> findByAutor(String nombreAutor) {
-
-		return repo.getLibrosAutor(nombreAutor);
-	}
-
-	public List<Libro> findByTitulo(String titulo) {
-		return repo.getLibroByTitulo(titulo);
-	}
+	List<Libro> findByTitulo(String titulo);
 }
