+++ src/main/resources/h2/test.trace.db
@@ -90,7 +90,7 @@ General error: "java.lang.IllegalStateException: The file is locked: nio:/home/d
 	at java.sql/java.sql.DriverManager.getConnection(DriverManager.java:229)
 	at com.kerjox.rubrica5bloque5.connection.AbstractConnection.open(AbstractConnection.java:16)
 	at com.kerjox.rubrica5bloque5.repos.AutoresRepo.insert(AutoresRepo.java:25)
-	at com.kerjox.rubrica5bloque5.services.AutoresService.insert(AutoresService.java:19)
+	at com.kerjox.rubrica5bloque5.services.AutoresServiceImpl.insert(AutoresService.java:19)
 	at com.kerjox.rubrica5bloque5.controllers.AutorController.create(AutorController.java:37)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:77)
@@ -202,7 +202,7 @@ General error: "java.lang.IllegalStateException: The file is locked: nio:/home/d
 	at java.sql/java.sql.DriverManager.getConnection(DriverManager.java:229)
 	at com.kerjox.rubrica5bloque5.connection.AbstractConnection.open(AbstractConnection.java:16)
 	at com.kerjox.rubrica5bloque5.repos.AutoresRepo.insert(AutoresRepo.java:25)
-	at com.kerjox.rubrica5bloque5.services.AutoresService.insert(AutoresService.java:19)
+	at com.kerjox.rubrica5bloque5.services.AutoresServiceImpl.insert(AutoresService.java:19)
 	at com.kerjox.rubrica5bloque5.controllers.AutorController.create(AutorController.java:37)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:77)
@@ -337,7 +337,7 @@ Column "ID" not found [42122-200]
 	at org.h2.jdbc.JdbcResultSet.getColumnIndex(JdbcResultSet.java:3195)
 	at org.h2.jdbc.JdbcResultSet.findColumn(JdbcResultSet.java:191)
 	at com.kerjox.rubrica5bloque5.repos.AutoresRepo.findAll(AutoresRepo.java:72)
-	at com.kerjox.rubrica5bloque5.services.AutoresService.findAll(AutoresService.java:31)
+	at com.kerjox.rubrica5bloque5.services.AutoresServiceImpl.findAll(AutoresService.java:31)
 	at com.kerjox.rubrica5bloque5.controllers.AutorController.list(AutorController.java:25)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:77)
@@ -493,7 +493,7 @@ General error: "java.lang.IllegalStateException: The file is locked: nio:/home/d
 	at java.sql/java.sql.DriverManager.getConnection(DriverManager.java:229)
 	at com.kerjox.rubrica5bloque5.connection.AbstractConnection.open(AbstractConnection.java:16)
 	at com.kerjox.rubrica5bloque5.repos.LibroRepo.findAll(LibroRepo.java:64)
-	at com.kerjox.rubrica5bloque5.services.LibroService.findAll(LibroService.java:27)
+	at com.kerjox.rubrica5bloque5.services.LibroServiceImpl.findAll(LibroService.java:27)
 	at com.kerjox.rubrica5bloque5.controllers.LibroController.list(LibroController.java:32)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:77)
@@ -584,7 +584,7 @@ General error: "java.lang.IllegalStateException: The file is locked: nio:/home/d
 	at java.sql/java.sql.DriverManager.getConnection(DriverManager.java:229)
 	at com.kerjox.rubrica5bloque5.connection.AbstractConnection.open(AbstractConnection.java:16)
 	at com.kerjox.rubrica5bloque5.repos.LibroRepo.findAll(LibroRepo.java:64)
-	at com.kerjox.rubrica5bloque5.services.LibroService.findAll(LibroService.java:27)
+	at com.kerjox.rubrica5bloque5.services.LibroServiceImpl.findAll(LibroService.java:27)
 	at com.kerjox.rubrica5bloque5.controllers.LibroController.list(LibroController.java:32)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:77)
@@ -675,7 +675,7 @@ General error: "java.lang.IllegalStateException: The file is locked: nio:/home/d
 	at java.sql/java.sql.DriverManager.getConnection(DriverManager.java:229)
 	at com.kerjox.rubrica5bloque5.connection.AbstractConnection.open(AbstractConnection.java:16)
 	at com.kerjox.rubrica5bloque5.repos.LibroRepo.findAll(LibroRepo.java:64)
-	at com.kerjox.rubrica5bloque5.services.LibroService.findAll(LibroService.java:27)
+	at com.kerjox.rubrica5bloque5.services.LibroServiceImpl.findAll(LibroService.java:27)
 	at com.kerjox.rubrica5bloque5.controllers.LibroController.list(LibroController.java:32)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:77)
